many ui improvements to vault2
passwords are now using scrollable list, instead of the paned
view from Precursor. Paned view doesn't make sense because the
panes are so small.

scrolling accelerates with time spent pushing the scroll wheel,
allowing lists of hundreds of passwords to be navigated in a few
seconds.

a bunch of unused APIs are now deprecated in the itemcache -
scrolling lists off-load most of the ui functions. itemcache is
now used just for caching entries from the PDDB and also, more
importantly, filtering entries - filters are of course TBD implementation